Adding Lateral flow
Some lines are added to accommodate the lateral flow into the code. Two new options of lateral flow components are added: a) Lateral flow as a time series (type 1) and b) lateral flow as function of upstream flow (type 2). An example is added in the repository showing a acse of Lower Mississippi from Baton Rouge to Head of Pass or South West Pass.